feat: 이미지 삭제 로직 생성
